+++ src/Opulence/Console/Commands/AboutCommand.php
@@ -32,12 +32,10 @@ EOF;
      * @param CommandCollection $commands The list of commands
      * @param PaddingFormatter $paddingFormatter The space padding formatter to use
      * @param string $applicationVersion The version number of the application
-     * @deprecated 1.1.0 The $applicationVersion parameter will soon not be accepted
      */
     public function __construct(
         CommandCollection $commands,
-        PaddingFormatter $paddingFormatter,
-        string $applicationVersion = ''
+        PaddingFormatter $paddingFormatter
     ) {
         parent::__construct();