+++ src/Opulence/Console/Kernel.php
@@ -17,7 +17,6 @@ use Opulence\Console\Commands\CommandCollection;
 use Opulence\Console\Commands\Compilers\ICompiler as ICommandCompiler;
 use Opulence\Console\Commands\HelpCommand;
 use Opulence\Console\Commands\ICommand;
-use Opulence\Console\Commands\VersionCommand;
 use Opulence\Console\Requests\IRequest;
 use Opulence\Console\Requests\Parsers\IParser;
 use Opulence\Console\Responses\Compilers\Compiler;
@@ -40,26 +39,21 @@ class Kernel
     private $commandCompiler = null;
     /** @var CommandCollection The list of commands to choose from */
     private $commandCollection = null;
-    /** @var string The version number of the application */
-    private $applicationVersion = 'Unknown';
 
     /**
      * @param IParser $requestParser The request parser to use
      * @param ICommandCompiler $commandCompiler The command compiler to use
      * @param CommandCollection $commandCollection The list of commands to choose from
      * @param string $applicationVersion The version number of the application
-     * @deprecated 1.1.0 The $applicationVersion parameter will soon not be accepted
      */
     public function __construct(
         IParser $requestParser,
         ICommandCompiler $commandCompiler,
-        CommandCollection $commandCollection,
-        string $applicationVersion = 'Unknown'
+        CommandCollection $commandCollection
     ) {
         $this->requestParser = $requestParser;
         $this->commandCompiler = $commandCompiler;
         $this->commandCollection = $commandCollection;
-        $this->applicationVersion = $applicationVersion;
     }
 
     /**
@@ -81,17 +75,13 @@ class Kernel
             if ($this->isInvokingHelpCommand($request)) {
                 // We are going to execute the help command
                 $compiledCommand = $this->getCompiledHelpCommand($request);
-            } elseif ($this->isInvokingVersionCommand($request)) {
-                // We are going to execute the version command
-                $compiledCommand = new VersionCommand($this->applicationVersion);
             } elseif ($this->commandCollection->has($request->getCommandName())) {
                 // We are going to execute the command that was entered
                 $command = $this->commandCollection->get($request->getCommandName());
                 $compiledCommand = $this->commandCompiler->compile($command, $request);
             } else {
                 // We are defaulting to the about command
-                $compiledCommand = new AboutCommand($this->commandCollection, new PaddingFormatter(),
-                    $this->applicationVersion);
+                $compiledCommand = new AboutCommand($this->commandCollection, new PaddingFormatter());
             }
 
             $statusCode = $compiledCommand->execute($response);
@@ -160,16 +150,5 @@ class Kernel
     private function isInvokingHelpCommand(IRequest $request) : bool
     {
         return $request->getCommandName() === 'help' || $request->optionIsSet('h') || $request->optionIsSet('help');
-    }
-
-    /**
-     * Gets whether or not the input is invoking the version command
-     *
-     * @param IRequest $request The parsed request
-     * @return bool True if it is invoking the version command, otherwise false
-     */
-    private function isInvokingVersionCommand(IRequest $request) : bool
-    {
-        return $request->optionIsSet('v') || $request->optionIsSet('version');
     }
 }